@@ -1,30 +1,33 @@
 {
-  "name": "Alg_visual",
+  "name": "alg-visual",
   "version": "1.0.0",
   "description": "",
   "main": "webpack.config.js",
   "scripts": {
-    "webpack": "./node_modules/.bin/webpack -w",
+    "webpack": "webpack --watch --mode development",
+    "build": "webpack --mode production",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "engines": {
-    "node": "6.3.0",
-    "npm": "3.10.3"
+    "node": ">=14.0.0",
+    "npm": ">=6.0.0"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel-core": "^6.17.0",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "css-loader": "^0.25.0",
-    "lodash": "^4.16.4",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-input-range": "^0.9.2",
-    "react-modal": "^1.5.2",
-    "webpack": "^1.13.2"
+    "@babel/core": "^7.23.0",
+    "@babel/preset-env": "^7.23.0",
+    "@babel/preset-react": "^7.23.0",
+    "babel-loader": "^9.1.0",
+    "css-loader": "^6.8.0",
+    "lodash": "^4.17.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-input-range": "^1.3.0",
+    "react-modal": "^3.16.0",
+    "style-loader": "^3.3.0",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.0"
   }
 }
